requirements.txt
django-quill-editor==0.1.40
Django==4.2.5